Fix type issues in controllers

+++ src/Controller/ControllerCart.php
@@ -16,28 +16,28 @@ $products = $modelCart->getProductsByIds($productIds);
 $cartCount = 0;
 $cartTotal = 0.0;
 foreach ($products as $product) {
-    $productId = (int) $product['id_article'];
-    $quantity = (int) ($cartItems[$productId] ?? 0);
+    $productId = $product['id_article'];
+    $quantity = ($cartItems[$productId] ?? 0);
     $cartCount += $quantity;
-    $cartTotal += (float)$product['prix_article'] * $quantity;
+    $cartTotal += $product['prix_article'] * $quantity;
 }
 
 $totalWithReduc = null;
 if (!empty($_SESSION['userid']) && !empty($products)) {
-    $reductionPercent = $modelCart->getMemberReductionPercent((int) $_SESSION['userid']);
+    $reductionPercent = $modelCart->getMemberReductionPercent($_SESSION['userid']);
 
     if ($reductionPercent !== null && $reductionPercent > 0) {
         $userReduction = 1 - ($reductionPercent / 100);
         $totalWithReduc = 0.0;
 
         foreach ($products as $product) {
-            $productId = (int) $product['id_article'];
-            $quantity = (int) ($cartItems[$productId] ?? 0);
+            $productId = $product['id_article'];
+            $quantity = ($cartItems[$productId] ?? 0);
 
             if (!empty($product['reduction_article'])) {
-                $totalWithReduc += (float)$product['prix_article'] * $quantity * $userReduction;
+                $totalWithReduc += $product['prix_article'] * $quantity * $userReduction;
             } else {
-                $totalWithReduc += (float)$product['prix_article'] * $quantity;
+                $totalWithReduc += $product['prix_article'] * $quantity;
             }
         }
     }
